requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.8.1
async-timeout==4.0.2
attrs==21.4.0
bcrypt==3.2.0
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.0
click==7.1.2
ecdsa==0.14.1
fastapi==0.63.0
fastapi-pagination==0.9.3
frozenlist==1.3.0
h11==0.13.0
idna==2.10
iniconfig==1.1.1
Mako==1.2.1
MarkupSafe==2.1.1
multidict==6.0.2
packaging==21.3
passlib==1.7.4
pluggy==0.13.1
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.7.3
pyparsing==3.0.9
pytest==6.2.1
python-dotenv==0.15.0
python-jose==3.2.0
python-multipart==0.0.5
requests==2.25.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.3.22
starlette==0.13.6
toml==0.10.2
urllib3==1.26.10
uvicorn==0.13.3
yarl==1.7.2
psycopg2-binary==2.9.3
